【提交摘要】升级框架 【详细内容】  - 优化：删除qf.BuildContext，调整到Context里面  - 优化：BaseDal的Delete方法，调整为仅返回error，当删除记录不存在时，会返回error
